modules/sponsorship_search: Move search_sponsor into it's own home #450
By it's own admission it's got complex logic, it's shared code so we
should have it somewhere with it's own test harness.
